refactor: 카테고리 - 파일 분리 (#01)
파일 분리되지 않은 클래스를 생성하였음.
